created NavBar cmpnt - refactored from layout
