+++ layouts/default.vue
@@ -59,25 +59,8 @@
           />
         </div>
 
-        <!-- Navigation Links -->
-        <ul class="text-inherit text-white">
-          <li v-for="section in routes" :key="section.title">
-            <nav-link :to="section.route">
-              {{ section.title }}
-            </nav-link>
-            <ul
-              v-if="section.subroutes"
-              v-show="useRoute().path.indexOf(section.route) != -1"
-              class="bg-slate-800/30 py-2"
-            >
-              <li v-for="page in section.subroutes" :key="page.key">
-                <nav-link :to="`${section.route}/${page.key}`" :indent="true">
-                  {{ page.name }}
-                </nav-link>
-              </li>
-            </ul>
-          </li>
-        </ul>
+        <!-- Sidebar Navigation -->
+        <NavBar />
 
         <!-- Report Issue UI -->
         <report-issue />
@@ -154,62 +137,9 @@ const { data: documents } = useDocuments({
   depth: 0,
 });
 
-const { data: classes } = useFindMany(API_ENDPOINTS.classes, {
-  fields: ['name', 'key'].join(),
-  is_subclass: false,
-});
-
 const no_available_sources = computed(() => documents.value?.length ?? 0);
 
 const isLoadingData = useIsFetching();
-
-const routes = computed(() => [
-  {
-    title: 'Classes',
-    route: '/classes',
-    subroutes: classes.value ?? [],
-  },
-  {
-    title: 'Races',
-    route: '/races',
-  },
-  {
-    title: 'Monsters',
-    route: '/monsters',
-  },
-  {
-    title: 'Magic Items',
-    route: '/magic-items',
-  },
-  {
-    title: 'Spells',
-    route: '/spells',
-  },
-  {
-    title: 'Backgrounds',
-    route: '/backgrounds',
-  },
-  {
-    title: 'Feats',
-    route: '/feats',
-  },
-  {
-    title: 'Equipment',
-    route: '/equipment',
-  },
-  {
-    title: 'Conditions',
-    route: '/conditions',
-  },
-  {
-    title: 'Rules',
-    route: '/rules',
-  },
-  {
-    title: 'API Docs',
-    route: '/api-docs',
-  },
-]);
 
 const router = useRouter();
 
@@ -218,13 +148,8 @@ function doSearch(searchText) {
   showSidebar.value = false;
 }
 
-function toggleSidebar() {
-  showSidebar.value = !showSidebar.value;
-}
-
-function hideSidebar() {
-  showSidebar.value = false;
-}
+const toggleSidebar = () => (showSidebar.value = !showSidebar.value);
+const hideSidebar = () => (showSidebar.value = false);
 </script>
 
 <style lang="scss">
